+++ opentelemetry-zipkin/src/exporter/mod.rs
@@ -2,12 +2,11 @@ mod env;
 mod model;
 mod uploader;
 
-use futures_core::future::BoxFuture;
 use http::Uri;
 use model::endpoint::Endpoint;
-use opentelemetry::trace::TraceError;
 use opentelemetry_http::HttpClient;
 use opentelemetry_sdk::error::OTelSdkResult;
+use opentelemetry_sdk::trace::TraceError;
 use opentelemetry_sdk::{trace, ExportError};
 use std::net::{AddrParseError, SocketAddr};
 use std::sync::Arc;
@@ -127,12 +126,8 @@ async fn zipkin_export(
 
 impl trace::SpanExporter for ZipkinExporter {
     /// Export spans to Zipkin collector.
-    fn export(&mut self, batch: Vec<trace::SpanData>) -> BoxFuture<'static, OTelSdkResult> {
-        Box::pin(zipkin_export(
-            batch,
-            self.uploader.clone(),
-            self.local_endpoint.clone(),
-        ))
+    async fn export(&mut self, batch: Vec<trace::SpanData>) -> OTelSdkResult {
+        zipkin_export(batch, self.uploader.clone(), self.local_endpoint.clone()).await
     }
 }
 
@@ -162,12 +157,6 @@ pub enum Error {
 }
 
 impl ExportError for Error {
-    fn exporter_name(&self) -> &'static str {
-        "zipkin"
-    }
-}
-
-impl opentelemetry::trace::ExportError for Error {
     fn exporter_name(&self) -> &'static str {
         "zipkin"
     }